moves services to scoped packages
