+++ internal/commands/export_test.go
@@ -6,8 +6,8 @@ import (
 	"github.com/patrickhuber/go-cross/platform"
 	"github.com/patrickhuber/go-di"
 	"github.com/patrickhuber/wrangle/internal/diff"
+	"github.com/patrickhuber/wrangle/internal/export"
 	"github.com/patrickhuber/wrangle/internal/host"
-	"github.com/patrickhuber/wrangle/internal/services"
 	"github.com/stretchr/testify/require"
 )
 
@@ -15,7 +15,7 @@ func TestExport(t *testing.T) {
 	h := host.NewTest(platform.Linux, nil, nil)
 	container := h.Container()
 
-	export, err := di.Resolve[services.Export](container)
+	export, err := di.Resolve[export.Service](container)
 	require.NoError(t, err)
 	require.NotNil(t, export)
