+++ internal/commands/hook_test.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/patrickhuber/go-cross/platform"
 	"github.com/patrickhuber/go-di"
+	"github.com/patrickhuber/wrangle/internal/hook"
 	"github.com/patrickhuber/wrangle/internal/host"
-	"github.com/patrickhuber/wrangle/internal/services"
 	"github.com/stretchr/testify/require"
 )
 
@@ -14,7 +14,7 @@ func TestHook(t *testing.T) {
 
 	s := host.NewTest(platform.Linux, nil, nil)
 	container := s.Container()
-	result, err := di.Resolve[services.Hook](container)
+	result, err := di.Resolve[hook.Service](container)
 
 	require.NoError(t, err)
 	require.NotNil(t, result)
